+++ lib/bonfire.dart
@@ -36,6 +36,8 @@ export 'package:bonfire/tiled/tiled_world_map.dart';
 export 'package:bonfire/util/background_color_game.dart';
 export 'package:bonfire/util/collision/collision.dart';
 export 'package:bonfire/util/direction.dart';
+export 'package:bonfire/util/direction_animations/simple_animation_enum.dart';
+export 'package:bonfire/util/direction_animations/simple_direction_animation.dart';
 export 'package:bonfire/util/extensions.dart';
 export 'package:bonfire/util/functions.dart';
 export 'package:bonfire/util/game_controller.dart';